requirements.txt
requests
discord.py
aiohttp
